backend/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
kagglehub>=0.3.0
